sql/catalog/lease: break lease.go into files
Too much code for one file.

Release note: None
