+++ pkg/sql/catalog/lease/lease.go
@@ -12,12 +12,8 @@
 package lease
 
 import (
-	"bytes"
 	"context"
 	"fmt"
-	"math/rand"
-	"sort"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -36,7 +32,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 	"github.com/cockroachdb/cockroach/pkg/sql/sqlutil"
-	"github.com/cockroachdb/cockroach/pkg/util/grpcutil"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/log/logcrash"
@@ -74,280 +69,6 @@ var LeaseJitterFraction = settings.RegisterFloatSetting(
 	"mean duration of sql descriptor leases, this actual duration is jitterred",
 	base.DefaultDescriptorLeaseJitterFraction,
 	between0and1inclusive)
-
-// LeaseRenewalDuration controls the default time before a lease expires when
-// acquisition to renew the lease begins.
-var LeaseRenewalDuration = settings.RegisterDurationSetting(
-	"sql.catalog.descriptor_lease_renewal_fraction",
-	"controls the default time before a lease expires when acquisition to renew the lease begins",
-	base.DefaultDescriptorLeaseRenewalTimeout)
-
-// A lease stored in system.lease.
-type storedLease struct {
-	id         descpb.ID
-	version    int
-	expiration tree.DTimestamp
-}
-
-func (s *storedLease) String() string {
-	return fmt.Sprintf("ID = %d ver=%d expiration=%s", s.id, s.version, s.expiration)
-}
-
-// descriptorVersionState holds the state for a descriptor version. This
-// includes the lease information for a descriptor version.
-// TODO(vivek): A node only needs to manage lease information on what it
-// thinks is the latest version for a descriptor.
-type descriptorVersionState struct {
-	t *descriptorState
-	// This descriptor is immutable and can be shared by many goroutines.
-	// Care must be taken to not modify it.
-	catalog.Descriptor
-
-	mu struct {
-		syncutil.Mutex
-
-		// The expiration time for the descriptor version. A transaction with
-		// timestamp T can use this descriptor version iff
-		// Descriptor.GetDescriptorModificationTime() <= T < expiration
-		//
-		// The expiration time is either the expiration time of the lease when a lease
-		// is associated with the version, or the ModificationTime of the next version
-		// when the version isn't associated with a lease.
-		expiration hlc.Timestamp
-
-		refcount int
-		// Set if the node has a lease on this descriptor version.
-		// Leases can only be held for the two latest versions of
-		// a descriptor. The latest version known to a node
-		// (can be different than the current latest version in the store)
-		// is always associated with a lease. The previous version known to
-		// a node might not necessarily be associated with a lease.
-		lease *storedLease
-	}
-}
-
-func (s *descriptorVersionState) Release(ctx context.Context) {
-	s.t.release(ctx, s)
-}
-
-func (s *descriptorVersionState) Underlying() catalog.Descriptor {
-	return s.Descriptor
-}
-
-func (s *descriptorVersionState) Expiration() hlc.Timestamp {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.mu.expiration
-}
-
-func (s *descriptorVersionState) SafeMessage() string {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return fmt.Sprintf("%d ver=%d:%s, refcount=%d", s.GetID(), s.GetVersion(), s.mu.expiration, s.mu.refcount)
-}
-
-func (s *descriptorVersionState) String() string {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.stringLocked()
-}
-
-// stringLocked reads mu.refcount and thus needs to have mu held.
-func (s *descriptorVersionState) stringLocked() string {
-	return fmt.Sprintf("%d(%q) ver=%d:%s, refcount=%d", s.GetID(), s.GetName(), s.GetVersion(), s.mu.expiration, s.mu.refcount)
-}
-
-// hasExpired checks if the descriptor is too old to be used (by a txn
-// operating) at the given timestamp.
-func (s *descriptorVersionState) hasExpired(timestamp hlc.Timestamp) bool {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.hasExpiredLocked(timestamp)
-}
-
-// hasExpired checks if the descriptor is too old to be used (by a txn
-// operating) at the given timestamp.
-func (s *descriptorVersionState) hasExpiredLocked(timestamp hlc.Timestamp) bool {
-	return s.mu.expiration.LessEq(timestamp)
-}
-
-func (s *descriptorVersionState) incRefcount() {
-	s.mu.Lock()
-	s.incRefcountLocked()
-	s.mu.Unlock()
-}
-
-func (s *descriptorVersionState) incRefcountLocked() {
-	s.mu.refcount++
-	if log.V(2) {
-		log.VEventf(context.TODO(), 2, "descriptorVersionState.incRef: %s", s.stringLocked())
-	}
-}
-
-func (s *descriptorVersionState) getExpiration() hlc.Timestamp {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	return s.mu.expiration
-}
-
-// The lease expiration stored in the database is of a different type.
-// We've decided that it's too much work to change the type to
-// hlc.Timestamp, so we're using this method to give us the stored
-// type: tree.DTimestamp.
-func storedLeaseExpiration(expiration hlc.Timestamp) tree.DTimestamp {
-	return tree.DTimestamp{Time: timeutil.Unix(0, expiration.WallTime).Round(time.Microsecond)}
-}
-
-// storage implements the operations for acquiring and releasing leases and
-// publishing a new version of a descriptor.
-// TODO (lucy,ajwerner): This could go in its own package and expose an API for
-// the manager. Some of these fields belong on the manager, in any case, since
-// they're only used by the manager and not by the store itself.
-type storage struct {
-	nodeIDContainer  *base.SQLIDContainer
-	db               *kv.DB
-	clock            *hlc.Clock
-	internalExecutor sqlutil.InternalExecutor
-	settings         *cluster.Settings
-	codec            keys.SQLCodec
-
-	// group is used for all calls made to acquireNodeLease to prevent
-	// concurrent lease acquisitions from the store.
-	group *singleflight.Group
-
-	outstandingLeases *metric.Gauge
-	testingKnobs      StorageTestingKnobs
-}
-
-// jitteredLeaseDuration returns a randomly jittered duration from the interval
-// [(1-leaseJitterFraction) * leaseDuration, (1+leaseJitterFraction) * leaseDuration].
-func (s storage) jitteredLeaseDuration() time.Duration {
-	leaseDuration := LeaseDuration.Get(&s.settings.SV)
-	jitterFraction := LeaseJitterFraction.Get(&s.settings.SV)
-	return time.Duration(float64(leaseDuration) * (1 - jitterFraction +
-		2*jitterFraction*rand.Float64()))
-}
-
-// acquire a lease on the most recent version of a descriptor. If the lease
-// cannot be obtained because the descriptor is in the process of being dropped
-// or offline (currently only applicable to tables), the error will be of type
-// inactiveTableError. The expiration time set for the lease > minExpiration.
-func (s storage) acquire(
-	ctx context.Context, minExpiration hlc.Timestamp, id descpb.ID,
-) (desc catalog.Descriptor, expiration hlc.Timestamp, _ error) {
-	err := s.db.Txn(ctx, func(ctx context.Context, txn *kv.Txn) (err error) {
-		// Run the descriptor read as high-priority, thereby pushing any intents out
-		// of its way. We don't want schema changes to prevent lease acquisitions;
-		// we'd rather force them to refresh. Also this prevents deadlocks in cases
-		// where the name resolution is triggered by the transaction doing the
-		// schema change itself.
-		if err := txn.SetUserPriority(roachpb.MaxUserPriority); err != nil {
-			return err
-		}
-		expiration = txn.ReadTimestamp().Add(int64(s.jitteredLeaseDuration()), 0)
-		if expiration.LessEq(minExpiration) {
-			// In the rare circumstances where expiration <= minExpiration
-			// use an expiration based on the minExpiration to guarantee
-			// a monotonically increasing expiration.
-			expiration = minExpiration.Add(int64(time.Millisecond), 0)
-		}
-
-		// TODO (lucy): Previously this called getTableDescFromID followed by a call
-		// to ValidateSelf() instead of Validate(), to avoid the cross-table
-		// checks. Does this actually matter? We already potentially do cross-table
-		// checks when populating pre-19.2 foreign keys.
-		desc, err = catalogkv.MustGetDescriptorByID(ctx, txn, s.codec, id)
-		if err != nil {
-			return err
-		}
-		if err := catalog.FilterDescriptorState(
-			desc, tree.CommonLookupFlags{IncludeOffline: true}, // filter dropped only
-		); err != nil {
-			return err
-		}
-		log.VEventf(ctx, 2, "storage acquired lease %v@%v", desc, expiration)
-		nodeID := s.nodeIDContainer.SQLInstanceID()
-		if nodeID == 0 {
-			panic("zero nodeID")
-		}
-
-		// We use string interpolation here, instead of passing the arguments to
-		// InternalExecutor.Exec() because we don't want to pay for preparing the
-		// statement (which would happen if we'd pass arguments). Besides the
-		// general cost of preparing, preparing this statement always requires a
-		// read from the database for the special descriptor of a system table
-		// (#23937).
-		ts := storedLeaseExpiration(expiration)
-		insertLease := fmt.Sprintf(
-			`INSERT INTO system.public.lease ("descID", version, "nodeID", expiration) VALUES (%d, %d, %d, %s)`,
-			desc.GetID(), desc.GetVersion(), nodeID, &ts,
-		)
-		count, err := s.internalExecutor.Exec(ctx, "lease-insert", txn, insertLease)
-		if err != nil {
-			return err
-		}
-		if count != 1 {
-			return errors.Errorf("%s: expected 1 result, found %d", insertLease, count)
-		}
-		s.outstandingLeases.Inc(1)
-		return nil
-	})
-	if err != nil {
-		return nil, hlc.Timestamp{}, err
-	}
-	if s.testingKnobs.LeaseAcquiredEvent != nil {
-		s.testingKnobs.LeaseAcquiredEvent(desc, err)
-	}
-	return desc, expiration, nil
-}
-
-// Release a previously acquired descriptor. Never let this method
-// read a descriptor because it can be called while modifying a
-// descriptor through a schema change before the schema change has committed
-// that can result in a deadlock.
-func (s storage) release(ctx context.Context, stopper *stop.Stopper, lease *storedLease) {
-	retryOptions := base.DefaultRetryOptions()
-	retryOptions.Closer = stopper.ShouldQuiesce()
-	firstAttempt := true
-	// This transaction is idempotent; the retry was put in place because of
-	// NodeUnavailableErrors.
-	for r := retry.Start(retryOptions); r.Next(); {
-		log.VEventf(ctx, 2, "storage releasing lease %+v", lease)
-		nodeID := s.nodeIDContainer.SQLInstanceID()
-		if nodeID == 0 {
-			panic("zero nodeID")
-		}
-		const deleteLease = `DELETE FROM system.public.lease ` +
-			`WHERE ("descID", version, "nodeID", expiration) = ($1, $2, $3, $4)`
-		count, err := s.internalExecutor.Exec(
-			ctx,
-			"lease-release",
-			nil, /* txn */
-			deleteLease,
-			lease.id, lease.version, nodeID, &lease.expiration,
-		)
-		if err != nil {
-			log.Warningf(ctx, "error releasing lease %q: %s", lease, err)
-			if grpcutil.IsConnectionRejected(err) {
-				return
-			}
-			firstAttempt = false
-			continue
-		}
-		// We allow count == 0 after the first attempt.
-		if count > 1 || (count == 0 && firstAttempt) {
-			log.Warningf(ctx, "unexpected results while deleting lease %+v: "+
-				"expected 1 result, found %d", lease, count)
-		}
-
-		s.outstandingLeases.Dec(1)
-		if s.testingKnobs.LeaseReleasedEvent != nil {
-			s.testingKnobs.LeaseReleasedEvent(
-				lease.id, descpb.DescriptorVersion(lease.version), err)
-		}
-		break
-	}
-}
 
 // WaitForNoVersion returns once there are no unexpired leases left
 // for any version of the descriptor.
@@ -438,212 +159,6 @@ func NewIDVersionPrev(name string, id descpb.ID, currVersion descpb.DescriptorVe
 	return IDVersion{Name: name, ID: id, Version: currVersion - 1}
 }
 
-// CountLeases returns the number of unexpired leases for a number of descriptors
-// each at a particular version at a particular time.
-func CountLeases(
-	ctx context.Context, executor sqlutil.InternalExecutor, versions []IDVersion, at hlc.Timestamp,
-) (int, error) {
-	var whereClauses []string
-	for _, t := range versions {
-		whereClauses = append(whereClauses,
-			fmt.Sprintf(`("descID" = %d AND version = %d AND expiration > $1)`,
-				t.ID, t.Version),
-		)
-	}
-
-	stmt := fmt.Sprintf(`SELECT count(1) FROM system.public.lease AS OF SYSTEM TIME '%s' WHERE `,
-		at.AsOfSystemTime()) +
-		strings.Join(whereClauses, " OR ")
-	values, err := executor.QueryRowEx(
-		ctx, "count-leases", nil, /* txn */
-		sessiondata.InternalExecutorOverride{User: security.RootUserName()},
-		stmt, at.GoTime(),
-	)
-	if err != nil {
-		return 0, err
-	}
-	if values == nil {
-		return 0, errors.New("failed to count leases")
-	}
-	count := int(tree.MustBeDInt(values[0]))
-	return count, nil
-}
-
-// Get the descriptor valid for the expiration time from the store.
-// We use a timestamp that is just less than the expiration time to read
-// a version of the descriptor. A descriptorVersionState with the
-// expiration time set to expiration is returned.
-//
-// This returns an error when Replica.checkTSAboveGCThresholdRLocked()
-// returns an error when the expiration timestamp is less than the storage
-// layer GC threshold.
-func (s storage) getForExpiration(
-	ctx context.Context, expiration hlc.Timestamp, id descpb.ID,
-) (catalog.Descriptor, error) {
-	var desc catalog.Descriptor
-	err := s.db.Txn(ctx, func(ctx context.Context, txn *kv.Txn) (err error) {
-		prevTimestamp := expiration.Prev()
-		txn.SetFixedTimestamp(ctx, prevTimestamp)
-		desc, err = catalogkv.MustGetDescriptorByID(ctx, txn, s.codec, id)
-		if err != nil {
-			return err
-		}
-		if prevTimestamp.LessEq(desc.GetModificationTime()) {
-			return errors.AssertionFailedf("unable to read descriptor"+
-				" (%d, %s) found descriptor with modificationTime %s",
-				id, expiration, desc.GetModificationTime())
-		}
-		// Create a descriptorVersionState with the descriptor and without a lease.
-		return nil
-	})
-	return desc, err
-}
-
-func (s storage) leaseRenewalTimeout() time.Duration {
-	return LeaseRenewalDuration.Get(&s.settings.SV)
-}
-
-// descriptorSet maintains an ordered set of descriptorVersionState objects
-// sorted by version. It supports addition and removal of elements, finding the
-// descriptor for a particular version, or finding the most recent version.
-// The order is maintained by insert and remove and there can only be a
-// unique entry for a version. Only the last two versions can be leased,
-// with the last one being the latest one which is always leased.
-//
-// Each entry represents a time span [ModificationTime, expiration)
-// and can be used by a transaction iif:
-// ModificationTime <= transaction.Timestamp < expiration.
-type descriptorSet struct {
-	data []*descriptorVersionState
-}
-
-func (l *descriptorSet) String() string {
-	var buf bytes.Buffer
-	for i, s := range l.data {
-		if i > 0 {
-			buf.WriteString(" ")
-		}
-		buf.WriteString(fmt.Sprintf("%d:%d", s.GetVersion(), s.getExpiration().WallTime))
-	}
-	return buf.String()
-}
-
-func (l *descriptorSet) insert(s *descriptorVersionState) {
-	i, match := l.findIndex(s.GetVersion())
-	if match {
-		panic("unable to insert duplicate lease")
-	}
-	if i == len(l.data) {
-		l.data = append(l.data, s)
-		return
-	}
-	l.data = append(l.data, nil)
-	copy(l.data[i+1:], l.data[i:])
-	l.data[i] = s
-}
-
-func (l *descriptorSet) remove(s *descriptorVersionState) {
-	i, match := l.findIndex(s.GetVersion())
-	if !match {
-		panic(errors.AssertionFailedf("can't find lease to remove: %s", s))
-	}
-	l.data = append(l.data[:i], l.data[i+1:]...)
-}
-
-func (l *descriptorSet) find(version descpb.DescriptorVersion) *descriptorVersionState {
-	if i, match := l.findIndex(version); match {
-		return l.data[i]
-	}
-	return nil
-}
-
-func (l *descriptorSet) findIndex(version descpb.DescriptorVersion) (int, bool) {
-	i := sort.Search(len(l.data), func(i int) bool {
-		s := l.data[i]
-		return s.GetVersion() >= version
-	})
-	if i < len(l.data) {
-		s := l.data[i]
-		if s.GetVersion() == version {
-			return i, true
-		}
-	}
-	return i, false
-}
-
-func (l *descriptorSet) findNewest() *descriptorVersionState {
-	if len(l.data) == 0 {
-		return nil
-	}
-	return l.data[len(l.data)-1]
-}
-
-func (l *descriptorSet) findVersion(version descpb.DescriptorVersion) *descriptorVersionState {
-	if len(l.data) == 0 {
-		return nil
-	}
-	// Find the index of the first lease with version > targetVersion.
-	i := sort.Search(len(l.data), func(i int) bool {
-		return l.data[i].GetVersion() > version
-	})
-	if i == 0 {
-		return nil
-	}
-	// i-1 is the index of the newest lease for the previous version (the version
-	// we're looking for).
-	s := l.data[i-1]
-	if s.GetVersion() == version {
-		return s
-	}
-	return nil
-}
-
-type descriptorState struct {
-	m       *Manager
-	id      descpb.ID
-	stopper *stop.Stopper
-
-	// renewalInProgress is an atomic indicator for when a renewal for a
-	// lease has begun. This is atomic to prevent multiple routines from
-	// entering renewal initialization.
-	renewalInProgress int32
-
-	mu struct {
-		syncutil.Mutex
-
-		// descriptors sorted by increasing version. This set always
-		// contains a descriptor version with a lease as the latest
-		// entry. There may be more than one active lease when the system is
-		// transitioning from one version of the descriptor to another or
-		// when the node preemptively acquires a new lease for a version
-		// when the old lease has not yet expired. In the latter case, a new
-		// entry is created with the expiration time of the new lease and
-		// the older entry is removed.
-		active descriptorSet
-
-		// Indicates that the descriptor has been, or is being, dropped or taken
-		// offline.
-		// If set, leases are released from the store as soon as their
-		// refcount drops to 0, as opposed to waiting until they expire.
-		// This flag will be unset by any subsequent lease acquisition, which can
-		// happen after the table came back online again after having been taken
-		// offline temporarily (as opposed to dropped).
-		takenOffline bool
-
-		// maxVersionSeen is used to prevent a race where a concurrent lease
-		// acquisition might miss an event indicating that there is a new version
-		// of a descriptor.
-		maxVersionSeen descpb.DescriptorVersion
-
-		// acquisitionsInProgress indicates that at least one caller is currently
-		// in the process of performing an acquisition. This tracking is critical
-		// to ensure that notifications of new versions which arrive before a lease
-		// acquisition finishes but indicate that that new lease is expired are not
-		// ignored.
-		acquisitionsInProgress int
-	}
-}
-
 // ensureVersion ensures that the latest version >= minVersion. It will
 // check if the latest known version meets the criterion, or attempt to
 // acquire a lease at the latest version with the hope that it meets
@@ -663,60 +178,6 @@ func ensureVersion(
 		return errors.Errorf("version %d for descriptor %s does not exist yet", minVersion, s.GetName())
 	}
 	return nil
-}
-
-// findForTimestamp finds a descriptor valid for the timestamp.
-// In the most common case the timestamp passed to this method is close
-// to the current time and in all likelihood the latest version of a
-// descriptor if valid is returned.
-//
-// This returns errRenewLease when there is no descriptor cached
-// or the latest descriptor version's ModificationTime satisfies the
-// timestamp while it's expiration time doesn't satisfy the timestamp.
-// This is an optimistic strategy betting that in all likelihood a
-// higher layer renewing the lease on the descriptor and populating
-// descriptorState will satisfy the timestamp on a subsequent call.
-//
-// In all other circumstances where a descriptor cannot be found for the
-// timestamp errOlderReadTableVersion is returned requesting a higher layer
-// to populate the descriptorState with a valid older version of the descriptor
-// before calling.
-//
-// The refcount for the returned descriptorVersionState is incremented.
-// It returns true if the descriptor returned is the known latest version
-// of the descriptor.
-func (t *descriptorState) findForTimestamp(
-	ctx context.Context, timestamp hlc.Timestamp,
-) (*descriptorVersionState, bool, error) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	// Acquire a lease if no descriptor exists in the cache.
-	if len(t.mu.active.data) == 0 {
-		return nil, false, errRenewLease
-	}
-
-	// Walk back the versions to find one that is valid for the timestamp.
-	for i := len(t.mu.active.data) - 1; i >= 0; i-- {
-		// Check to see if the ModificationTime is valid.
-		if desc := t.mu.active.data[i]; desc.GetModificationTime().LessEq(timestamp) {
-			latest := i+1 == len(t.mu.active.data)
-			if !desc.hasExpired(timestamp) {
-				// Existing valid descriptor version.
-				desc.incRefcount()
-				return desc, latest, nil
-			}
-
-			if latest {
-				// Renew the lease if the lease has expired
-				// The latest descriptor always has a lease.
-				return nil, false, errRenewLease
-			}
-			break
-		}
-	}
-
-	return nil, false, errReadOlderVersion
 }
 
 type historicalDescriptor struct {
@@ -860,95 +321,6 @@ func (m *Manager) AcquireFreshestFromStore(ctx context.Context, id descpb.ID) er
 	return nil
 }
 
-var _ redact.SafeMessager = (*descriptorVersionState)(nil)
-
-// upsertLeaseLocked inserts a lease for a particular descriptor version.
-// If an existing lease exists for the descriptor version it replaces
-// it and returns it.
-func (t *descriptorState) upsertLeaseLocked(
-	ctx context.Context, desc catalog.Descriptor, expiration hlc.Timestamp,
-) (createdDescriptorVersionState *descriptorVersionState, toRelease *storedLease, _ error) {
-	if t.mu.maxVersionSeen < desc.GetVersion() {
-		t.mu.maxVersionSeen = desc.GetVersion()
-	}
-	s := t.mu.active.find(desc.GetVersion())
-	if s == nil {
-		if t.mu.active.findNewest() != nil {
-			log.Infof(ctx, "new lease: %s", desc)
-		}
-		descState := newDescriptorVersionState(t, desc, expiration, true /* isLease */)
-		t.mu.active.insert(descState)
-		return descState, nil, nil
-	}
-
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	// The desc is replacing an existing one at the same version.
-	if !s.mu.expiration.Less(expiration) {
-		// This is a violation of an invariant and can actually not
-		// happen. We return an error here to aid in further investigations.
-		return nil, nil, errors.AssertionFailedf("lease expiration monotonicity violation, (%s) vs (%s)", s, desc)
-	}
-
-	// subsume the refcount of the older lease. This is permitted because
-	// the new lease has a greater expiration than the older lease and
-	// any transaction using the older lease can safely use a deadline set
-	// to the older lease's expiration even though the older lease is
-	// released! This is because the new lease is valid at the same desc
-	// version at a greater expiration.
-	s.mu.expiration = expiration
-	toRelease = s.mu.lease
-	s.mu.lease = &storedLease{
-		id:         desc.GetID(),
-		version:    int(desc.GetVersion()),
-		expiration: storedLeaseExpiration(expiration),
-	}
-	if log.ExpensiveLogEnabled(ctx, 2) {
-		log.VEventf(ctx, 2, "replaced lease: %s with %s", toRelease, s.mu.lease)
-	}
-	return nil, toRelease, nil
-}
-
-func newDescriptorVersionState(
-	t *descriptorState, desc catalog.Descriptor, expiration hlc.Timestamp, isLease bool,
-) *descriptorVersionState {
-	descState := &descriptorVersionState{
-		t:          t,
-		Descriptor: desc,
-	}
-	descState.mu.expiration = expiration
-	if isLease {
-		descState.mu.lease = &storedLease{
-			id:         desc.GetID(),
-			version:    int(desc.GetVersion()),
-			expiration: storedLeaseExpiration(expiration),
-		}
-	}
-	return descState
-}
-
-// removeInactiveVersions removes inactive versions in t.mu.active.data with
-// refcount 0. t.mu must be locked. It returns leases that need to be released.
-func (t *descriptorState) removeInactiveVersions() []*storedLease {
-	var leases []*storedLease
-	// A copy of t.mu.active.data must be made since t.mu.active.data will be changed
-	// within the loop.
-	for _, desc := range append([]*descriptorVersionState(nil), t.mu.active.data...) {
-		func() {
-			desc.mu.Lock()
-			defer desc.mu.Unlock()
-			if desc.mu.refcount == 0 {
-				t.mu.active.remove(desc)
-				if l := desc.mu.lease; l != nil {
-					desc.mu.lease = nil
-					leases = append(leases, l)
-				}
-			}
-		}()
-	}
-	return leases
-}
-
 // If the lease cannot be obtained because the descriptor is in the process of
 // being dropped or offline, the error will be of type inactiveTableError.
 // The boolean returned is true if this call was actually responsible for the
@@ -1000,56 +372,6 @@ func acquireNodeLease(ctx context.Context, m *Manager, id descpb.ID) (bool, erro
 		}
 	}
 	return didAcquire, nil
-}
-
-// release returns a descriptorVersionState that needs to be released from
-// the store.
-func (t *descriptorState) release(ctx context.Context, s *descriptorVersionState) {
-
-	// Decrements the refcount and returns true if the lease has to be removed
-	// from the store.
-	decRefcount := func(s *descriptorVersionState) *storedLease {
-		// Figure out if we'd like to remove the lease from the store asap (i.e.
-		// when the refcount drops to 0). If so, we'll need to mark the lease as
-		// invalid.
-		removeOnceDereferenced := t.m.removeOnceDereferenced() ||
-			// Release from the store if the descriptor has been dropped or taken
-			// offline.
-			t.mu.takenOffline ||
-			// Release from the store if the lease is not for the latest
-			// version; only leases for the latest version can be acquired.
-			s != t.mu.active.findNewest() ||
-			s.GetVersion() < t.mu.maxVersionSeen
-
-		s.mu.Lock()
-		defer s.mu.Unlock()
-		s.mu.refcount--
-		if log.V(2) {
-			log.VEventf(context.TODO(), 2, "release: %s", s.stringLocked())
-		}
-		if s.mu.refcount < 0 {
-			panic(errors.AssertionFailedf("negative ref count: %s", s))
-		}
-
-		if s.mu.refcount == 0 && s.mu.lease != nil && removeOnceDereferenced {
-			l := s.mu.lease
-			s.mu.lease = nil
-			return l
-		}
-		return nil
-	}
-	maybeRemoveLease := func() *storedLease {
-		t.mu.Lock()
-		defer t.mu.Unlock()
-		if l := decRefcount(s); l != nil {
-			t.mu.active.remove(s)
-			return l
-		}
-		return nil
-	}
-	if l := maybeRemoveLease(); l != nil {
-		releaseLease(l, t.m)
-	}
 }
 
 // releaseLease from store.
@@ -1136,57 +458,6 @@ func purgeOldVersions(
 	return err
 }
 
-// maybeQueueLeaseRenewal queues a lease renewal if there is not already a lease
-// renewal in progress.
-func (t *descriptorState) maybeQueueLeaseRenewal(
-	ctx context.Context, m *Manager, id descpb.ID, name string,
-) error {
-	if !atomic.CompareAndSwapInt32(&t.renewalInProgress, 0, 1) {
-		return nil
-	}
-
-	// Start the renewal. When it finishes, it will reset t.renewalInProgress.
-	return t.stopper.RunAsyncTask(context.Background(),
-		"lease renewal", func(ctx context.Context) {
-			t.startLeaseRenewal(ctx, m, id, name)
-		})
-}
-
-// startLeaseRenewal starts a singleflight.Group to acquire a lease.
-// This function blocks until lease acquisition completes.
-// t.renewalInProgress must be set to 1 before calling.
-func (t *descriptorState) startLeaseRenewal(
-	ctx context.Context, m *Manager, id descpb.ID, name string,
-) {
-	log.VEventf(ctx, 1,
-		"background lease renewal beginning for id=%d name=%q",
-		id, name)
-	if _, err := acquireNodeLease(ctx, m, id); err != nil {
-		log.Errorf(ctx,
-			"background lease renewal for id=%d name=%q failed: %s",
-			id, name, err)
-	} else {
-		log.VEventf(ctx, 1,
-			"background lease renewal finished for id=%d name=%q",
-			id, name)
-	}
-	atomic.StoreInt32(&t.renewalInProgress, 0)
-}
-
-// markAcquisitionStart increments the acquisitionsInProgress counter.
-func (t *descriptorState) markAcquisitionStart(ctx context.Context) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-	t.mu.acquisitionsInProgress++
-}
-
-// markAcquisitionDone decrements the acquisitionsInProgress counter.
-func (t *descriptorState) markAcquisitionDone(ctx context.Context) {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-	t.mu.acquisitionsInProgress--
-}
-
 // AcquireBlockType is the type of blocking result event when
 // calling LeaseAcquireResultBlockEvent.
 type AcquireBlockType int
@@ -1199,164 +470,6 @@ const (
 	// from descriptorState.acquireFreshestFromStore().
 	AcquireFreshestBlock
 )
-
-// StorageTestingKnobs contains testing knobs.
-type StorageTestingKnobs struct {
-	// Called after a lease is removed from the store, with any operation error.
-	// See LeaseRemovalTracker.
-	LeaseReleasedEvent func(id descpb.ID, version descpb.DescriptorVersion, err error)
-	// Called after a lease is acquired, with any operation error.
-	LeaseAcquiredEvent func(desc catalog.Descriptor, err error)
-	// Called before waiting on a results from a DoChan call of acquireNodeLease
-	// in descriptorState.acquire() and descriptorState.acquireFreshestFromStore().
-	LeaseAcquireResultBlockEvent func(leaseBlockType AcquireBlockType)
-	// RemoveOnceDereferenced forces leases to be removed
-	// as soon as they are dereferenced.
-	RemoveOnceDereferenced bool
-}
-
-// ModuleTestingKnobs is part of the base.ModuleTestingKnobs interface.
-func (*StorageTestingKnobs) ModuleTestingKnobs() {}
-
-var _ base.ModuleTestingKnobs = &StorageTestingKnobs{}
-
-// ManagerTestingKnobs contains test knobs.
-type ManagerTestingKnobs struct {
-
-	// A callback called after the leases are refreshed as a result of a gossip update.
-	TestingDescriptorRefreshedEvent func(descriptor *descpb.Descriptor)
-
-	// TestingDescriptorUpdateEvent is a callback when an update is received, before
-	// the leases are refreshed. If a non-nil error is returned, the update is
-	// ignored.
-	TestingDescriptorUpdateEvent func(descriptor *descpb.Descriptor) error
-
-	// To disable the deletion of orphaned leases at server startup.
-	DisableDeleteOrphanedLeases bool
-
-	// VersionPollIntervalForRangefeeds controls the polling interval for the
-	// check whether the requisite version for rangefeed-based notifications has
-	// been finalized.
-	//
-	// TODO(ajwerner): Remove this and replace it with a callback.
-	VersionPollIntervalForRangefeeds time.Duration
-
-	LeaseStoreTestingKnobs StorageTestingKnobs
-}
-
-var _ base.ModuleTestingKnobs = &ManagerTestingKnobs{}
-
-// ModuleTestingKnobs is part of the base.ModuleTestingKnobs interface.
-func (*ManagerTestingKnobs) ModuleTestingKnobs() {}
-
-// nameCacheKey is a key for the descriptor cache, with the same fields
-// as the system.namespace key: name is the descriptor name; parentID is
-// populated for schemas, descriptors, and types; and parentSchemaID is
-// populated for descriptors and types.
-type nameCacheKey struct {
-	parentID       descpb.ID
-	parentSchemaID descpb.ID
-	name           string
-}
-
-// nameCache is a cache of descriptor name -> latest version mappings.
-// The Manager updates the cache every time a lease is acquired or released
-// from the store. The cache maintains the latest version for each name.
-// All methods are thread-safe.
-type nameCache struct {
-	mu          syncutil.Mutex
-	descriptors map[nameCacheKey]*descriptorVersionState
-}
-
-// Resolves a (qualified) name to the descriptor's ID.
-// Returns a valid descriptorVersionState for descriptor with that name,
-// if the name had been previously cached and the cache has a descriptor
-// version that has not expired. Returns nil otherwise.
-// This method handles normalizing the descriptor name.
-// The descriptor's refcount is incremented before returning, so the caller
-// is responsible for releasing it to the leaseManager.
-func (c *nameCache) get(
-	parentID descpb.ID, parentSchemaID descpb.ID, name string, timestamp hlc.Timestamp,
-) *descriptorVersionState {
-	c.mu.Lock()
-	desc, ok := c.descriptors[makeNameCacheKey(parentID, parentSchemaID, name)]
-	c.mu.Unlock()
-	if !ok {
-		return nil
-	}
-	desc.mu.Lock()
-	if desc.mu.lease == nil {
-		desc.mu.Unlock()
-		// This get() raced with a release operation. Remove this cache
-		// entry if needed.
-		c.remove(desc)
-		return nil
-	}
-
-	defer desc.mu.Unlock()
-
-	if !NameMatchesDescriptor(desc, parentID, parentSchemaID, name) {
-		panic(errors.AssertionFailedf("out of sync entry in the name cache. "+
-			"Cache entry: (%d, %d, %q) -> %d. Lease: (%d, %d, %q).",
-			parentID, parentSchemaID, name,
-			desc.GetID(),
-			desc.GetParentID(), desc.GetParentSchemaID(), desc.GetName()),
-		)
-	}
-
-	// Expired descriptor. Don't hand it out.
-	if desc.hasExpiredLocked(timestamp) {
-		return nil
-	}
-
-	desc.incRefcountLocked()
-	return desc
-}
-
-func (c *nameCache) insert(desc *descriptorVersionState) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	key := makeNameCacheKey(desc.GetParentID(), desc.GetParentSchemaID(), desc.GetName())
-	existing, ok := c.descriptors[key]
-	if !ok {
-		c.descriptors[key] = desc
-		return
-	}
-	// If we already have a lease in the cache for this name, see if this one is
-	// better (higher version or later expiration).
-	if desc.GetVersion() > existing.GetVersion() ||
-		(desc.GetVersion() == existing.GetVersion() &&
-			existing.getExpiration().Less(desc.getExpiration())) {
-		// Overwrite the old lease. The new one is better. From now on, we want
-		// clients to use the new one.
-		c.descriptors[key] = desc
-	}
-}
-
-func (c *nameCache) remove(desc *descriptorVersionState) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-
-	key := makeNameCacheKey(desc.GetParentID(), desc.GetParentSchemaID(), desc.GetName())
-	existing, ok := c.descriptors[key]
-	if !ok {
-		// Descriptor for lease not found in name cache. This can happen if we had
-		// a more recent lease on the descriptor in the nameCache, then the
-		// descriptor gets dropped, then the more recent lease is remove()d - which
-		// clears the cache.
-		return
-	}
-	// If this was the lease that the cache had for the descriptor name, remove
-	// it. If the cache had some other descriptor, this remove is a no-op.
-	if existing == desc {
-		delete(c.descriptors, key)
-	}
-}
-
-func makeNameCacheKey(parentID descpb.ID, parentSchemaID descpb.ID, name string) nameCacheKey {
-	return nameCacheKey{parentID, parentSchemaID, name}
-}
 
 // Manager manages acquiring and releasing per-descriptor leases. It also
 // handles resolving descriptor names to descriptor IDs. The leases are managed
@@ -1510,7 +623,7 @@ func (m *Manager) AcquireByName(
 		return desc, nil
 	}
 	// Check if we have cached an ID for this name.
-	descVersion := m.names.get(parentID, parentSchemaID, name, timestamp)
+	descVersion := m.names.get(ctx, parentID, parentSchemaID, name, timestamp)
 	if descVersion != nil {
 		if descVersion.GetModificationTime().LessEq(timestamp) {
 			expiration := descVersion.getExpiration()
@@ -2067,29 +1180,4 @@ func (m *Manager) VisitLeases(
 			return
 		}
 	}
-}
-
-// TestingAcquireAndAssertMinVersion acquires a read lease for the specified
-// ID. The lease is grabbed on the latest version if >= specified version.
-// It returns a descriptor and an expiration time valid for the timestamp.
-// This method is useful for testing and is only intended to be used in that
-// context.
-func (m *Manager) TestingAcquireAndAssertMinVersion(
-	ctx context.Context, timestamp hlc.Timestamp, id descpb.ID, minVersion descpb.DescriptorVersion,
-) (LeasedDescriptor, error) {
-	t := m.findDescriptorState(id, true)
-	if err := ensureVersion(ctx, id, minVersion, m); err != nil {
-		return nil, err
-	}
-	desc, _, err := t.findForTimestamp(ctx, timestamp)
-	if err != nil {
-		return nil, err
-	}
-	return desc, nil
-}
-
-// TestingOutstandingLeasesGauge returns the outstanding leases gauge that is
-// used by this lease manager.
-func (m *Manager) TestingOutstandingLeasesGauge() *metric.Gauge {
-	return m.storage.outstandingLeases
 }
