+++ pkg/sql/catalog/lease/lease_internal_test.go
@@ -380,6 +380,7 @@ CREATE TEMP TABLE t2 (temp int);
 	for _, tableName := range []string{"t", "t2"} {
 		tableDesc := catalogkv.TestingGetTableDescriptor(kvDB, keys.SystemSQLCodec, "defaultdb", tableName)
 		lease := leaseManager.names.get(
+			context.Background(),
 			tableDesc.GetParentID(),
 			descpb.ID(keys.PublicSchemaID),
 			tableName,
@@ -419,11 +420,23 @@ CREATE TABLE t.test (k CHAR PRIMARY KEY, v CHAR);
 	}
 
 	// Check that the cache has been updated.
-	if leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), "test", s.Clock().Now()) != nil {
+	if leaseManager.names.get(
+		context.Background(),
+		tableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		"test",
+		s.Clock().Now(),
+	) != nil {
 		t.Fatalf("old name still in cache")
 	}
 
-	lease := leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), "test2", s.Clock().Now())
+	lease := leaseManager.names.get(
+		context.Background(),
+		tableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		"test2",
+		s.Clock().Now(),
+	)
 	if lease == nil {
 		t.Fatalf("new name not found in cache")
 	}
@@ -444,11 +457,23 @@ CREATE TABLE t.test (k CHAR PRIMARY KEY, v CHAR);
 	}
 
 	// Check that the cache has been updated.
-	if leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), "test2", s.Clock().Now()) != nil {
+	if leaseManager.names.get(
+		context.Background(),
+		tableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		"test2",
+		s.Clock().Now(),
+	) != nil {
 		t.Fatalf("old name still in cache")
 	}
 
-	lease = leaseManager.names.get(newTableDesc.GetParentID(), tableDesc.GetParentSchemaID(), "test2", s.Clock().Now())
+	lease = leaseManager.names.get(
+		context.Background(),
+		newTableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		"test2",
+		s.Clock().Now(),
+	)
 	if lease == nil {
 		t.Fatalf("new name not found in cache")
 	}
@@ -483,7 +508,13 @@ CREATE TABLE t.%s (k CHAR PRIMARY KEY, v CHAR);
 
 	// Check the assumptions this tests makes: that there is a cache entry
 	// (with a valid lease).
-	if lease := leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), tableName, s.Clock().Now()); lease == nil {
+	if lease := leaseManager.names.get(
+		context.Background(),
+		tableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		tableName,
+		s.Clock().Now(),
+	); lease == nil {
 		t.Fatalf("name cache has no unexpired entry for (%d, %s)", tableDesc.GetParentID(), tableName)
 	} else {
 		lease.Release(context.Background())
@@ -492,7 +523,13 @@ CREATE TABLE t.%s (k CHAR PRIMARY KEY, v CHAR);
 	leaseManager.ExpireLeases(s.Clock())
 
 	// Check the name no longer resolves.
-	if lease := leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), tableName, s.Clock().Now()); lease != nil {
+	if lease := leaseManager.names.get(
+		context.Background(),
+		tableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		tableName,
+		s.Clock().Now(),
+	); lease != nil {
 		t.Fatalf("name cache has unexpired entry for (%d, %s): %s", tableDesc.GetParentID(), tableName, lease)
 	}
 }
@@ -530,7 +567,13 @@ CREATE TABLE t.%s (k CHAR PRIMARY KEY, v CHAR);
 	}
 
 	// There is a cache entry.
-	lease := leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), tableName, s.Clock().Now())
+	lease := leaseManager.names.get(
+		context.Background(),
+		tableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		tableName,
+		s.Clock().Now(),
+	)
 	if lease == nil {
 		t.Fatalf("name cache has no unexpired entry for (%d, %s)", tableDesc.GetParentID(), tableName)
 	}
@@ -543,7 +586,7 @@ CREATE TABLE t.%s (k CHAR PRIMARY KEY, v CHAR);
 	}
 
 	// Check the name resolves to the new lease.
-	newLease := leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), tableName, s.Clock().Now())
+	newLease := leaseManager.names.get(context.Background(), tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), tableName, s.Clock().Now())
 	if newLease == nil {
 		t.Fatalf("name cache doesn't contain entry for (%d, %s)", tableDesc.GetParentID(), tableName)
 	}
@@ -584,7 +627,13 @@ CREATE TABLE t.test (k CHAR PRIMARY KEY, v CHAR);
 	tableDesc := catalogkv.TestingGetTableDescriptor(kvDB, keys.SystemSQLCodec, "t", "test")
 
 	// Check that we cannot get the table by a different name.
-	if leaseManager.names.get(tableDesc.GetParentID(), tableDesc.GetParentSchemaID(), "tEsT", s.Clock().Now()) != nil {
+	if leaseManager.names.get(
+		context.Background(),
+		tableDesc.GetParentID(),
+		tableDesc.GetParentSchemaID(),
+		"tEsT",
+		s.Clock().Now(),
+	) != nil {
 		t.Fatalf("lease manager incorrectly found table with different case")
 	}
 }
